Float -> double

+++ ch03/3.4/weatherenrich/src/main/java/weatherenrich/events/EnrichedEvent.java
@@ -5,7 +5,7 @@ public class EnrichedEvent extends RawEvent {
   public final EnrichedContext context;
 
   public EnrichedEvent(RawEvent rawEvent,
-    Float temperature, String[] conditions) {                          // a
+    double temperature, String[] conditions) {                         // a
 
     this.subject = rawEvent.subject;
     this.verb = rawEvent.verb;
@@ -19,7 +19,7 @@ public class EnrichedEvent extends RawEvent {
     public final Weather weather;
 
     public EnrichedContext(String timestamp,
-      Float temperature, String[] conditions) {
+      double temperature, String[] conditions) {
 
       this.timestamp = timestamp;
       this.weather = new Weather(temperature, conditions);
@@ -29,7 +29,7 @@ public class EnrichedEvent extends RawEvent {
       public final double temperature;
       public final String[] conditions;
 
-      public Weather(Float temperature, String[] conditions) {
+      public Weather(double temperature, String[] conditions) {
         this.temperature = temperature;
         this.conditions = conditions;
       }
